// src/data/hero-crop-selections.generated.ts
/** Generated by scripts/apply-hero-crops.mjs. Do not edit by hand. */
import type { HeroCropRecord } from '@/lib/hero-crop/types'

export const heroCropSelectionsGenerated: Readonly<Record<string, HeroCropRecord>> = {
  "thumb-1920-1381117.webp": {
    "filename": "thumb-1920-1381117.webp",
    "desktop": {
      "x": 44.186046511627815,
      "y": 52.593912534100376,
      "zoom": 1
    },
    "mobile": {
      "x": 15.641025641025777,
      "y": 46.560064498790666,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:11:51.281Z"
  },
  "43935854_p0_master1200.webp": {
    "filename": "43935854_p0_master1200.webp",
    "desktop": {
      "x": 47.84980954836947,
      "y": 31.753134057347985,
      "zoom": 1
    },
    "mobile": {
      "x": 50,
      "y": 50,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:12:36.632Z"
  },
  "thumb-1920-949729.webp": {
    "filename": "thumb-1920-949729.webp",
    "desktop": {
      "x": 45.788688525976205,
      "y": 15.652953408551816,
      "zoom": 1
    },
    "mobile": {
      "x": 82.82051282051283,
      "y": 14.645107348681748,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:13:09.509Z"
  },
  "thumb-1920-725406.webp": {
    "filename": "thumb-1920-725406.webp",
    "desktop": {
      "x": 44.91572356823063,
      "y": 30.23255052188368,
      "zoom": 1
    },
    "mobile": {
      "x": 72.56410256410244,
      "y": 50.00000000000002,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:28:08.520Z"
  },
  "thumb-1920-986446.webp": {
    "filename": "thumb-1920-986446.webp",
    "desktop": {
      "x": 33.42793382175734,
      "y": 49.91054909263443,
      "zoom": 1
    },
    "mobile": {
      "x": 50,
      "y": 50,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:14:15.047Z"
  },
  "thumb-1920-556375.webp": {
    "filename": "thumb-1920-556375.webp",
    "desktop": {
      "x": 35.321828377926565,
      "y": 49.463321852390024,
      "zoom": 1
    },
    "mobile": {
      "x": 50,
      "y": 50,
      "zoom": 1
    },
    "schemaVersion": 1,
    "updatedAt": "2026-09-01T07:14:45.742Z"
  }
}